ErrorList, request -> command structure of program, volunteer's ability to add with it's photos and delete pet's photos.

+++ PetFamily.Backend/src/PetFamily.Application/Volunteers/Delete/HardDeleteVolunteerHandler.cs
@@ -1,5 +1,7 @@
 using CSharpFunctionalExtensions;
+using FluentValidation;
 using Microsoft.Extensions.Logging;
+using PetFamily.Application.Extensions;
 using PetFamily.Application.Volunteers.UpdateMainInfo;
 using PetFamily.Domain.PetContext.ValueObjects.VolunteerVO;
 using PetFamily.Domain.Shared.CustomErrors;
@@ -10,24 +12,31 @@ public class HardDeleteVolunteerHandler
 {
     private readonly IVolunteersRepository _volunteersRepository;
     private readonly ILogger<HardDeleteVolunteerHandler> _logger;
+    private readonly IValidator<DeleteVolunteerCommand> _validator;
 
     public HardDeleteVolunteerHandler(
         IVolunteersRepository volunteersRepository,
-        ILogger<HardDeleteVolunteerHandler> logger)
+        ILogger<HardDeleteVolunteerHandler> logger,
+        IValidator<DeleteVolunteerCommand> validator)
     {
         _volunteersRepository = volunteersRepository;
         _logger = logger;
+        _validator = validator;
     }
 
-    public async Task<Result<Guid, Error>> HandleAsync(
-        DeleteVolunteerRequest request,
+    public async Task<Result<Guid, ErrorList>> HandleAsync(
+        DeleteVolunteerCommand command,
         CancellationToken cancellationToken)
     {
-        var volunteerId = VolunteerId.Create(request.Id);
+        var validationResult = await _validator.ValidateAsync(command, cancellationToken);
+        if (validationResult.IsValid == false)
+            return validationResult.ToErrorList();
+        
+        var volunteerId = VolunteerId.Create(command.Id);
 
         var volunteerResult = await _volunteersRepository.GetByIdAsync(volunteerId, cancellationToken);
         if (volunteerResult.IsFailure)
-            return volunteerResult.Error;
+            return volunteerResult.Error.ToErrorList();
         
         await _volunteersRepository.DeleteAsync(volunteerResult.Value, cancellationToken);
